% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rorpy.R
\name{rorpy}
\alias{rorpy}
\title{R or Python?  Classification of webpages by code content.}
\usage{
rorpy(url)
}
\arguments{
\item{A}{character vector of length n containing urls.}
}
\value{
A tibble containing the probability that the input url contains R 
code (column \code{r}), Python code (column \code{py}) or another code type 
(column \code{other}).  If no code can be found the probability vector will 
be 0.  If there are problems fetching data from a url, the NAs will be returned 
instead of classification probabilities.
}
\description{
This function fetches code chunks from a vector of urls (where 
code is assumed to be tagged \code{<pre>}, \code{<code>} or \code{<textarea>}).   
The function uses a pretrained random forest classifier to calculates the 
probability that the code is R, Python or neither.
}
\examples{
# not run:
# rorpy("https://google.com") # no code here...
# rorpy("http://dplyr.tidyverse.org") # 99\\\% sure it's R.......  
# rorpy("https://keras.io") # also about 99\\\% sure it's python
}
